WIP
Enable ALSA driver and System V IPC in Linux Kernel. For
ALSA, the debug is enabled and will be removed once
this commit is ready.

Add description of descriptor chaining, yet need refactoring
the description as well as code (exists uncertainty of
the query struct).

Succeed to initialize virtio-snd.

Handle requests in control and TX queue, and print the address and length
of each virtq element to check the validness of self-implement
queue.

For macOS, an experimental core audio library check is implemented.

As the driver sends the PCM frames asynchronously, use a dedicated
thread for receiving frames from driver.

The reason that pcm_release state hangs is because of PulseAudio,
need to address this later.

Add ring buffer struct so that we don't need to allocate
space frequently.

Can play guest sound. However, the sound is lagging.